Optimizacion de formatos de las vistas y recyclersView
